+++ packages/app/src/hooks/use-archive-agent.test.ts
@@ -1,19 +1,16 @@
-/**
- * @vitest-environment jsdom
- */
-import React from "react";
 import type { DaemonClient } from "@getpaseo/client/internal/daemon-client";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { act, renderHook, waitFor } from "@testing-library/react";
+import { QueryClient } from "@tanstack/react-query";
 import { beforeEach, describe, expect, it } from "vitest";
 import type { Agent } from "@/stores/session-store";
 import { useSessionStore } from "@/stores/session-store";
 import { agentHistoryQueryKey } from "./agent-history-query-key";
+import { applyArchivedAgentCloseResults } from "./use-archive-agent";
 import {
-  __private__,
-  applyArchivedAgentCloseResults,
-  usePendingArchiveAgentIds,
-} from "./use-archive-agent";
+  isAgentArchiving,
+  removeAgentFromListPayload,
+  selectPendingArchiveAgentIds,
+  setAgentArchiving,
+} from "./use-archive-agent.pure";
 
 function makeAgent(overrides: Partial<Agent> = {}): Agent {
   return {
@@ -56,14 +53,14 @@ describe("useArchiveAgent", () => {
     const queryClient = new QueryClient();
 
     expect(
-      __private__.isAgentArchiving({
+      isAgentArchiving({
         queryClient,
         serverId: "server-a",
         agentId: "agent-1",
       }),
     ).toBe(false);
 
-    __private__.setAgentArchiving({
+    setAgentArchiving({
       queryClient,
       serverId: "server-a",
       agentId: "agent-1",
@@ -71,21 +68,21 @@ describe("useArchiveAgent", () => {
     });
 
     expect(
-      __private__.isAgentArchiving({
+      isAgentArchiving({
         queryClient,
         serverId: "server-a",
         agentId: "agent-1",
       }),
     ).toBe(true);
     expect(
-      __private__.isAgentArchiving({
+      isAgentArchiving({
         queryClient,
         serverId: "server-a",
         agentId: "agent-2",
       }),
     ).toBe(false);
 
-    __private__.setAgentArchiving({
+    setAgentArchiving({
       queryClient,
       serverId: "server-a",
       agentId: "agent-1",
@@ -93,7 +90,7 @@ describe("useArchiveAgent", () => {
     });
 
     expect(
-      __private__.isAgentArchiving({
+      isAgentArchiving({
         queryClient,
         serverId: "server-a",
         agentId: "agent-1",
@@ -102,7 +99,7 @@ describe("useArchiveAgent", () => {
   });
 
   it("selects pending archive ids for a single server", () => {
-    const pendingIds = __private__.selectPendingArchiveAgentIds(
+    const pendingIds = selectPendingArchiveAgentIds(
       {
         "server-a:agent-1": true,
         "server-a:agent-2": true,
@@ -114,42 +111,13 @@ describe("useArchiveAgent", () => {
     expect(Array.from(pendingIds)).toEqual(["agent-1", "agent-2"]);
   });
 
-  it("subscribes to pending archive ids through react-query", async () => {
-    const queryClient = new QueryClient();
-    const wrapper = ({ children }: { children: React.ReactNode }) =>
-      React.createElement(QueryClientProvider, { client: queryClient }, children);
-
-    const { result } = renderHook(() => usePendingArchiveAgentIds("server-a"), { wrapper });
-
-    expect(Array.from(result.current)).toEqual([]);
-
-    act(() => {
-      __private__.setAgentArchiving({
-        queryClient,
-        serverId: "server-a",
-        agentId: "agent-1",
-        isArchiving: true,
-      });
-      __private__.setAgentArchiving({
-        queryClient,
-        serverId: "server-b",
-        agentId: "agent-2",
-        isArchiving: true,
-      });
-    });
-
-    await waitFor(() => {
-      expect(Array.from(result.current)).toEqual(["agent-1"]);
-    });
-  });
-
   it("removes an archived agent from cached list payloads", () => {
     const payload = {
       entries: [{ agent: { id: "agent-1" } }, { agent: { id: "agent-2" } }],
       pageInfo: { hasMore: false },
     };
 
-    const next = __private__.removeAgentFromListPayload(payload, "agent-1");
+    const next = removeAgentFromListPayload(payload, "agent-1");
 
     expect(next.entries).toEqual([{ agent: { id: "agent-2" } }]);
     expect(next.pageInfo).toEqual({ hasMore: false });
